Bugs (#646)
* bugs

* mutation change
