+++ web/rainmaker/dev-packages/egov-pt-dev/src/ui-config/screens/specs/pt-mutation/my-applications.js
@@ -4,8 +4,7 @@ import { getCommonHeader } from "egov-ui-framework/ui-config/screens/specs/utils
 const header = getCommonHeader(
   {
     labelName: "My Applications",
-    labelKey: "PT_MUTATION_MY_APPLICATIONS",
-    dynamicArray: [0]
+    labelKey: "PT_MUTATION_MY_APPLICATION_HEADER",
   },
   {
     classes: {
