+++ web/rainmaker/dev-packages/egov-ui-kit-dev/src/common/common/PTHeader/index.scss
@@ -11,8 +11,8 @@
             div{
                 button{
                     display: block !important;
-                    height: 70px !important;
-                    width: 70px !important;
+                    height: 65px !important;
+                    width: 157px !important;
                 }
             }
         }
